fix union query and IBeneficiaryActivities
